// frontend/src/lib/api/api.interfaces.d.ts
export interface Organization {
    id: string;
    name: string;
    logo: string | null;
    role: OrganizationRoleEnum;
    created_at: Date;
    updated_at: Date;
    members_count: number;
    permissions: OrganizationPermissions;
}
// ...
export interface OrganizationPermissions {
    organization: [
        PermissionEnum.read,
        PermissionEnum.update,
        PermissionEnum.delete,
        PermissionEnum.leave
    ];
    organization_user: Omit<PermissionEnum, PermissionEnum.deploy>;
    organization_invite: Omit<PermissionEnum, PermissionEnum.update & PermissionEnum.deploy>;
    billing: [PermissionEnum.read, PermissionEnum.update];
    orbit: [PermissionEnum.create];
}
export interface OrbitPermissions {
    orbit: Omit<PermissionEnum, PermissionEnum.deploy>;
    orbit_user: Omit<PermissionEnum, PermissionEnum.deploy>;
    artifact: PermissionEnum;
    collection: PermissionEnum;
    track: Omit<PermissionEnum, PermissionEnum.deploy>;
}
export declare enum PermissionEnum {
    list = "list",
    read = "read",
    create = "create",
    update = "update",
    delete = "delete",
    deploy = "deploy",
    leave = "leave"
}